refactoring: structural simplification
